// bm/www_src/app/gf/types.ts
export interface TimerData {
    active: boolean;
    remaining_minutes: number;
    remaining_seconds: number;
    total_start_time: number;
}